requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.3
certifi==2025.4.26
charset-normalizer==3.4.2
cleo==2.1.0
click==8.2.0
cohere==5.15.0
colorama==0.4.6
crashtest==0.4.1
distlib==0.3.9
distro==1.9.0
dulwich==0.22.8
faiss-cpu==1.11.0
fastavro==1.10.0
fastjsonschema==2.21.1
filelock==3.18.0
findpython==0.6.3
Flask==3.1.1
Flask-Pydantic==0.13.1
fsspec==2025.3.2
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.2
idna==3.10
iniconfig==2.1.0
installer==0.7.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
keyring==25.6.0
langchain==0.3.25
langchain-core==0.3.59
langchain-openai==0.3.16
langchain-text-splitters==0.3.8
langsmith==0.3.42
MarkupSafe==3.0.2
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
numpy==2.2.5
openai==1.78.1
orjson==3.10.18
packaging==24.2
pandas==2.2.3
pbs-installer==2025.4.9
pillow==11.2.1
pkginfo==1.12.1.2
platformdirs==4.3.8
pluggy==1.6.0
poetry==2.1.3
poetry-core==2.1.3
pydantic==2.11.4
pydantic_core==2.33.2
pyproject_hooks==1.2.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.4.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
state==0.1.1.dev0
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
trove-classifiers==2025.5.9.12
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.31.2
Werkzeug==3.1.3
zstandard==0.23.0
